First working bytecode example.
