+++ core/src/cpp/test.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
-gcc -Wall -W -Werror interpreter.cpp dummy.cpp -o dummy
+gcc -Wall -W -Werror interpreter.cpp interpreter_ea.cpp interpreter_run.cpp dummy.cpp -o dummy
 ./dummy
 rm dummy
